refactor: split TxOrderCard in separate files
